Moved covariate types to a more appropriate package.
