+++ .idea/csv-editor.xml
@@ -10,13 +10,6 @@
             </Attribute>
           </value>
         </entry>
-        <entry key="/punit-junit5/src/test/resources/input/test-inputs.csv">
-          <value>
-            <Attribute>
-              <option name="separator" value="," />
-            </Attribute>
-          </value>
-        </entry>
       </map>
     </option>
   </component>
