+++ punit-core/src/main/java/org/javai/punit/api/covariate/CovariateDeclaration.java
@@ -5,7 +5,6 @@ import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import org.javai.punit.api.CovariateCategory;
 import org.javai.punit.api.ServiceContractAttributes;
 import org.javai.punit.internal.util.HashUtils;
