+++ punit-core/src/test/java/org/javai/punit/api/covariate/CovariateDeclarationTest.java
@@ -8,7 +8,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import org.javai.punit.api.CovariateCategory;
 import org.javai.punit.api.ServiceContractAttributes;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
